General improvements to the code

+++ app/Listeners/AddCompletedDevice.php
@@ -25,28 +25,26 @@ class AddCompletedDevice
     public function handle(DeleteDevice $event): void
     {
         $device = Device::find($event->id);
-        if ($device) {
-            if ($device->deliver_to_client === true) {
-                $client = $device->client;
-                if ($client) {
-                    $completedDevice = CompletedDevice::create([
-                        'model' => $device->model,
-                        'imei' => $device->imei,
-                        'client_id' => $device->client_id,
-                        'user_id' => $device->user_id,
-                        'info' => $device->info,
-                        'problem' => $device->problem,
-                        'cost' => $device->cost,
-                        'status' => $device->status,
-                        'fix_steps' => $device->fix_steps,
-                        'date_receipt' => $device->date_receipt,
-                    ]);
-                    if ($completedDevice) {
-                        $client->decrement('devices_count');
-                    }
-                    if ($device->deliver_to_customer === true) {
-                        $device->delete();
-                    }
+        if ($device && $device->deliver_to_client === true) {
+            $client = $device->client;
+            if ($client) {
+                $completedDevice = CompletedDevice::create([
+                    'model' => $device->model,
+                    'imei' => $device->imei,
+                    'client_id' => $device->client_id,
+                    'user_id' => $device->user_id,
+                    'info' => $device->info,
+                    'problem' => $device->problem,
+                    'cost' => $device->cost,
+                    'status' => $device->status,
+                    'fix_steps' => $device->fix_steps,
+                    'date_receipt' => $device->date_receipt,
+                ]);
+                if ($completedDevice) {
+                    $client->decrement('devices_count');
+                }
+                if ($device->deliver_to_customer === true) {
+                    $device->delete();
                 }
             }
         }
